requirements.txt
# Fort Cordis - Dependências para deploy (Streamlit Community Cloud / produção)
# Python 3.10+ recomendado

streamlit>=1.28.0,<2.0
pandas>=2.0
beautifulsoup4>=4.12
fpdf2>=2.7.0
Pillow>=10.0
